Extracted wasm to standalone controller
