+++ web/landing/templates/playground/index.html.twig
@@ -13,25 +13,30 @@
 
 {% block main %}
 <div id="playground"
-     data-controller="playground"
-     data-playground-php-js-value="{{ asset('wasm/php.js') }}"
-     data-playground-php-wasm-value="{{ asset('wasm/php.wasm') }}"
-     data-playground-flow-phar-value="{{ asset('wasm/flow.phar') }}">
+     data-controller="playground-editor wasm"
+     data-playground-editor-code-editor-outlet="#code-editor"
+     data-wasm-php-js-value="{{ asset('wasm/php.js') }}"
+     data-wasm-php-wasm-value="{{ asset('wasm/php.wasm') }}"
+     data-wasm-flow-phar-value="{{ asset('wasm/flow.phar') }}"
+     data-action="wasm:ready->playground-editor#onWasmReady
+                  wasm:progress->playground-editor#onWasmProgress
+                  wasm:output->playground-editor#onWasmOutput
+                  wasm:error->playground-editor#onWasmError">
     <h1>Playground</h1>
 
     <div id="loader">
-        <div data-playground-target="loadingMessage"></div>
+        <div data-playground-editor-target="loadingMessage"></div>
         <div class="loading-progress">
-            <div data-playground-target="loadingBar" class="loading-bar"></div>
+            <div data-playground-editor-target="loadingBar" class="loading-bar"></div>
         </div>
-        <div data-playground-target="loadingPercent"></div>
+        <div data-playground-editor-target="loadingPercent"></div>
     </div>
 
-    <div class="navigation" data-playground-target="navigation">
+    <div class="navigation" data-playground-editor-target="navigation">
         <div class="actions">
             <button class="hover:text-green-600"
-                    data-playground-target="runButton"
-                    data-action="click->playground#run">
+                    data-playground-editor-target="runButton"
+                    data-action="click->playground-editor#run">
                 Run <img src="{{ asset('images/icons/play.svg') }}"
                      alt="Run"
                      class="ml-1 inline-block"
@@ -41,9 +46,9 @@
             </button>
         </div>
     </div>
-    <div class="editor" data-playground-target="editor">
+    <div class="editor" data-playground-editor-target="editor">
         <div class="code-container">
-                <textarea id="features_example" data-controller="code-editor" data-theme="ace/theme/github_dark">
+                <textarea id="code-editor" data-controller="code-editor" data-theme="ace/theme/github_dark">
 <?php
 
 use function Flow\ETL\DSL\{data_frame, from_array, to_output, ref, lit};
@@ -64,9 +69,9 @@ data_frame()
     ->run();</textarea>
         </div>
     </div>
-    <div class="output-container" data-playground-target="outputContainer">
+    <div class="output-container" data-playground-editor-target="outputContainer">
         <h2>Output</h2>
-        <pre id="output" data-playground-target="output"></pre>
+        <pre id="output" data-playground-editor-target="output"></pre>
     </div>
 </div>
 {% endblock %}
